vrclient: Unwrap array portion of Vulkan depth texture too
No Man's Sky does Submit_TextureWithDepth with an invalid depth
texture handle. This has gone unnoticed as the SteamVR compositor
doesn't currently use the depth texture. As it is unused, for the
case of just Submit_TextureWithDepth, set the texture depth handle
to NULL instead of unpacking. This avoids crashing No Man's Sky.